#98: Move everything regarding the `module` command in role_management
